Refactor clinical generators into lifecycle module
